After chatting with LaMont Jones about making my "upstream" version numbers play nicest with Debian packaging, I decided to move to this system:
version.py specifies a TARGET and a VERSION.  The TARGET is the version number that we hope to see when we fully release.  The VERSION number is the code name of the current project aiming toward TARGET.

So when the VERSION and TARGET match, it's a release and that's our public "version number".  But when they don't we take advantage of the magic earlier-than-nothing debian '~' separator to sling on our VERSION after the TARGET.  

So this is similar to the 1.0~alpha -> 1.0~beta -> 1.0~rc1 -> 1.0~rc2 -> 1.0 progress, only we use cuter names from the get-go.

This could be total crack, because we become locked into particular TARGETs.  This should force us to be conservative at the start, at the very least.